refactor: ajustando o nome da implementação do EmailMessage e atualizando scripts
