+++ package.json
@@ -19,9 +19,9 @@
     "test:watch": "jest --watch --runInBand --config ./jest.unit.config.ts",
     "test:cov": "jest --coverage --runInBand --config ./jest.unit.config.ts",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:unit": "jest --runInBand --config ./jest.unit.config.ts",
-    "test:int": "jest --runInBand --config ./jest.int.config.ts",
-    "test:e2e": "jest --runInBand --config ./jest.e2e.config.ts",
+    "test:unit": "dotenv -e .env.development -- jest --runInBand --config ./jest.unit.config.ts",
+    "test:int": "dotenv -e .env.development -- jest --runInBand --config ./jest.int.config.ts",
+    "test:e2e": "dotenv -e .env.development -- jest --runInBand --config ./jest.e2e.config.ts",
     "docker:build": "docker compose build",
     "docker:up": "docker compose up && npm run prisma:seed",
     "docker:build-up": "docker compose up --build && npm run prisma:seed",
